Extract projection
